仓储部分单据删除判断

+++ ktg-mes/src/main/java/com/ktg/mes/wm/controller/WmIssueHeaderController.java
@@ -123,6 +123,11 @@ public class WmIssueHeaderController extends BaseController
     {
         for (long issueId:issueIds
              ) {
+            WmIssueHeader header = wmIssueHeaderService.selectWmIssueHeaderByIssueId(issueId);
+            if(!UserConstants.ORDER_STATUS_PREPARE.equals(header.getStatus())){
+                return AjaxResult.error("只能删除草稿状态的单据!");
+            }
+
             wmIssueLineService.deleteByIssueHeaderId(issueId);
         }
